+++ sources/CONSOLE_SESSIONS.sql
@@ -6,10 +6,10 @@ begin
     dbms_output.put_line('- Table CONSOLE_SESSIONS not found, run creation command');
     execute immediate q'{
       create table console_sessions (
+        client_identifier varchar2 (64 byte)  not null  ,
         init_by           varchar2 (64 byte)            ,
         init_sysdate      date                not null  ,
         exit_sysdate      date                not null  ,
-        client_identifier varchar2 (64 byte)  not null  ,
         level_id          number   ( 1,0)     not null  ,
         level_name        varchar2 (10 byte)  not null  ,
         cache_size        number   ( 4,0)     not null  ,
@@ -21,12 +21,16 @@ begin
         console_env       varchar2 ( 1 byte)  not null  ,
         --
         constraint  console_sessions_pk   primary key ( client_identifier          )  ,
-        constraint  console_sessions_ck1  check       ( level_id    in (1,2,3,4,5) )  ,
-        constraint  console_sessions_ck2  check       ( call_stack  in ('Y','N')   )  ,
-        constraint  console_sessions_ck3  check       ( user_env    in ('Y','N')   )  ,
-        constraint  console_sessions_ck4  check       ( apex_env    in ('Y','N')   )  ,
-        constraint  console_sessions_ck5  check       ( cgi_env     in ('Y','N')   )  ,
-        constraint  console_sessions_ck6  check       ( console_env in ('Y','N')   )
+        constraint  console_sessions_ck1  check       ( call_stack  in ('Y','N')   )  ,
+        constraint  console_sessions_ck2  check       ( user_env    in ('Y','N')   )  ,
+        constraint  console_sessions_ck3  check       ( apex_env    in ('Y','N')   )  ,
+        constraint  console_sessions_ck4  check       ( cgi_env     in ('Y','N')   )  ,
+        constraint  console_sessions_ck5  check       ( console_env in ('Y','N')   )  ,
+        --
+        constraint  console_sessions_ck6  check       ( level_id   in (1, 2, 3, 4, 5)                                                          )  ,
+        constraint  console_sessions_ck7  check       ( level_name =  decode(level_id, 1,'error', 2,'warning', 3,'info', 4,'debug', 5,'trace') )  ,
+        --
+        constraint  console_sessions_ck8  check       ( check_interval between 1 and 60 )
       ) organization index
     }';
   else
@@ -40,10 +44,10 @@ comment on table  console_sessions                     is 'Holds the sessions th
 comment on column console_sessions.init_by             is 'The user who initiated the logging.';
 comment on column console_sessions.init_sysdate        is 'The logging start date for the nominated client identifier.';
 comment on column console_sessions.exit_sysdate        is 'The planned logging end date for the nominated client identifier.';
-comment on column console_sessions.client_identifier   is 'The client identifier provided by the application or console itself.';
-comment on column console_sessions.level_id            is 'The defined log level ID. Any session not listed here has the default log level of 1 (error).';
-comment on column console_sessions.level_name          is 'The defined log level name. Any session not listed here has the default log level of Error.';
-comment on column console_sessions.check_interval is 'The number of seconds a session in logging mode looks for a changed configuration. Defaults to 10.';
+comment on column console_sessions.client_identifier   is 'The client identifier provided by the application or console itself (this is the primary key).';
+comment on column console_sessions.level_id            is 'The defined log level ID. Any session not listed here has the configured global log level defined in CONSOLE_CONF.';
+comment on column console_sessions.level_name          is 'The defined log level name.';
+comment on column console_sessions.check_interval      is 'The number of seconds a session looks for a changed configuration. Defaults to 10.';
 comment on column console_sessions.cache_size          is 'The number of log entries to cache before they are written down into the log table. Errors are flushing always the cache. If greater then zero and no errors occur you can loose log entries in shared environments like APEX.';
 comment on column console_sessions.call_stack          is 'Should the call_stack be included.';
 comment on column console_sessions.user_env            is 'Should the user environment be included.';